Refactored into folders and fixed issue when tapping plus button
