+++ alarm_ranges/ios/Flutter/flutter_export_environment.sh
@@ -1,0 +1,10 @@
+#!/bin/sh
+# This is a generated file; do not edit or check into version control.
+export "FLUTTER_ROOT=/Users/mho/workspace/flutter"
+export "FLUTTER_APPLICATION_PATH=/Users/mho/Documents/flutter-alarm-ranges/alarm_ranges"
+export "FLUTTER_TARGET=/Users/mho/Documents/flutter-alarm-ranges/alarm_ranges/lib/main.dart"
+export "FLUTTER_BUILD_DIR=build"
+export "SYMROOT=${SOURCE_ROOT}/../build/ios"
+export "FLUTTER_FRAMEWORK_DIR=/Users/mho/workspace/flutter/bin/cache/artifacts/engine/ios"
+export "FLUTTER_BUILD_NAME=1.0.0"
+export "FLUTTER_BUILD_NUMBER=1"
